user checker

+++ apps/api/views/user.py
@@ -22,6 +22,8 @@ from apps.core.models import OfferChat, Message
 from apps.core.models.user import User
 from apps.api.serializers.user import UserSerializer
 
+from object_checker.base_object_checker import has_object_permission
+
 
 def _get_user(request, user_id: UUID) -> User:
     try:
@@ -74,7 +76,8 @@ class UserDetail(SecuredView):
             raise ValidationException(request, form)
 
         user = _get_user(request, user_id)
-
+        if not has_object_permission('check_user', user=request.user, obj=user):
+            raise ProblemDetailException(request, _('Permission denied.'), status=HTTPStatus.FORBIDDEN)
 
         if User.objects.filter(email=form.cleaned_data['email']).exclude(pk=user.id).exists():
             raise ProblemDetailException(
@@ -89,7 +92,8 @@ class UserDetail(SecuredView):
     @transaction.atomic
     def delete(self, request, user_id: UUID):
         user = _get_user(request, user_id)
-
+        if not has_object_permission('check_user', user=request.user, obj=user):
+            raise ProblemDetailException(request, _('Permission denied.'), status=HTTPStatus.FORBIDDEN)
         user.delete()
 
         return SingleResponse(request)
@@ -118,6 +122,8 @@ class UserProfileImage(SecuredView):
             raise ValidationException(request, form)
 
         user = _get_user(request, user_id)
+        if not has_object_permission('check_user', user=request.user, obj=user):
+            raise ProblemDetailException(request, _('Permission denied.'), status=HTTPStatus.FORBIDDEN)
 
         user.image.save(name=f"{uuid.uuid4()}{mimetypes.guess_extension(form.cleaned_data['image'].content_type)}",
                         content=form.cleaned_data['image'])
@@ -128,6 +134,8 @@ class UserProfileImage(SecuredView):
     def delete(self, request, user_id: UUID):
 
         user = _get_user(request, user_id)
+        if not has_object_permission('check_user', user=request.user, obj=user):
+            raise ProblemDetailException(request, _('Permission denied.'), status=HTTPStatus.FORBIDDEN)
 
         if user.image.path == settings.DEFAULT_IMAGE:
             raise ProblemDetailException(
@@ -144,6 +152,9 @@ class UserProfileImage(SecuredView):
 class UserChat(SecuredView):
 
     def get(self, request, user_id: UUID):
+
+        if not has_object_permission('check_user_chats', user=request.user, obj=user_id):
+            raise ProblemDetailException(request, _('Permission denied.'), status=HTTPStatus.FORBIDDEN)
 
         last_message = Message.objects.filter(offer_chat=OuterRef('pk')).order_by('-created_at')
